+++ CHANGELOG.md
@@ -9,6 +9,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 - `PMG\Queue\Driver\PheanstalkDriver` no longer accepts arrays as its `$options`
   argument, instead pass `PMG\Queue\Driver\Pheanstalk\ArrayOptions`.
+- `PMG\Queue\Driver\Pheanstalk\BuryFailureStrategy` no longer accepts an integer
+  argument to its constructor, pass an options instance instead.
 
 ## 5.2.0